requirements.txt
essentia-tensorflow
transformers
diffusers
numpy==1.26.4  # Numpy < 2.0 needed for Essentia
librosa
matplotlib
torchsde

# !wget https://essentia.upf.edu/models/music-style-classification/discogs-effnet/discogs-effnet-bs64-1.pb
# !wget https://essentia.upf.edu/models/classification-heads/mtg_jamendo_moodtheme/mtg_jamendo_moodtheme-discogs-effnet-1.pb
# !wget https://essentia.upf.edu/models/classification-heads/deam/deam-audioset-vggish-2.pb 
# !wget https://essentia.upf.edu/models/feature-extractors/vggish/audioset-vggish-3.pb